backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.118.3
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
oauthlib==3.3.1
openai==2.6.1
orjson==3.11.3
pycparser==2.23
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
sentry-sdk==2.41.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
tweepy==4.16.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
praw
Pillow
